RDA: Move `DefUseState` in its own file
